+++ KeepingYouAwake/KYAAdvancedPreferencesViewController.h
@@ -9,6 +9,15 @@
 #import <Cocoa/Cocoa.h>
 
 @interface KYAAdvancedPreferencesViewController : NSViewController
+
+/**
+ *  Determines if the current Mac has a built-in battery.
+ */
+@property (nonatomic, readonly, getter=isBatteryStatusAvailable) BOOL batteryStatusAvailable;
+
+/**
+ *  A table view reference.
+ */
 @property (weak, nonatomic, nullable) IBOutlet NSTableView *tableView;
 
 /**
@@ -17,5 +26,13 @@
  *  @param sender A sender.
  */
 - (IBAction)resetAdvancedPreferences:(nullable id)sender;
+
+/**
+ *  Is responsible to notifiy the app controller about any
+ *  battery status preference changes.
+ *
+ *  @param sender A sender.
+ */
+- (IBAction)batteryStatusPreferencesChanged:(nullable id)sender;
 
 @end