+++ KeepingYouAwake/KYAAppController.m
@@ -10,10 +10,16 @@
 #import "KYASleepWakeTimer.h"
 #import "KYAEventHandler.h"
 #import "KYAMenuBarIcon.h"
+#import "KYABatteryStatus.h"
+#import "KYABatteryCapacityThreshold.h"
 #import "NSUserDefaults+Keys.h"
 
 @interface KYAAppController () <NSUserNotificationCenterDelegate>
-@property (strong, nonatomic, readwrite) KYASleepWakeTimer *sleepWakeTimer;
+@property (nonatomic, readwrite) KYASleepWakeTimer *sleepWakeTimer;
+
+// Battery Status
+@property (nonatomic) KYABatteryStatus *batteryStatus;
+@property (nonatomic, getter=isBatteryOverrideEnabled) BOOL batteryOverrideEnabled;
 
 // Status Item
 @property (strong, nonatomic) NSStatusItem *statusItem;
@@ -49,6 +55,7 @@
                                                    object:nil
          ];
         
+        [self configureBatteryStatus];
         [self configureEventHandler];
     }
     return self;
@@ -56,7 +63,9 @@
 
 - (void)dealloc
 {
-    [[NSNotificationCenter defaultCenter] removeObserver:self name:NSApplicationDidFinishLaunchingNotification object:nil];
+    NSNotificationCenter *notificationCenter = [NSNotificationCenter defaultCenter];
+    [notificationCenter removeObserver:self name:NSApplicationDidFinishLaunchingNotification object:nil];
+    [notificationCenter removeObserver:self name:kKYABatteryCapacityThresholdDidChangeNotification object:nil];
 }
 
 - (void)awakeFromNib
@@ -181,6 +190,13 @@
         return;
     }
     
+    // Check battery overrides and register for capacity changes.
+    [self checkAndEnableBatteryOverride];
+    if([[NSUserDefaults standardUserDefaults] kya_isBatteryCapacityThresholdEnabled])
+    {
+        [self.batteryStatus registerForCapacityChangesIfNeeded];
+    }
+    
     [self.sleepWakeTimer scheduleWithTimeInterval:timeInterval completion:^(BOOL cancelled) {
         // Post notifications
         if([[NSUserDefaults standardUserDefaults] kya_areNotificationsEnabled])
@@ -216,17 +232,20 @@
 
 - (void)terminateTimer
 {
-    [self.sleepWakeTimer invalidate];
+    [self disableBatteryOverride];
+    [self.batteryStatus unregisterFromCapacityChanges];
+    
+    if([self.sleepWakeTimer isScheduled])
+    {
+        [self.sleepWakeTimer invalidate];
+    }
 }
 
 #pragma mark - Menu Delegate
 
 - (IBAction)selectTimeInterval:(NSMenuItem *)sender
 {
-    if([self.sleepWakeTimer isScheduled])
-    {
-        [self.sleepWakeTimer invalidate];
-    }
+    [self terminateTimer];
 	
 	if (sender.alternate)
 	{
@@ -307,6 +326,76 @@
     return YES;
 }
 
+#pragma mark - Battery Status
+
+- (KYABatteryStatus *)batteryStatus
+{
+    if(_batteryStatus == nil)
+    {
+        _batteryStatus = [KYABatteryStatus new];
+        
+        __weak typeof(self) weakSelf = self;
+        _batteryStatus.capacityChangeHandler = ^(CGFloat capacity) {
+            [weakSelf batteryCapacityDidChange:capacity];
+        };
+    }
+    return _batteryStatus;
+}
+
+- (void)configureBatteryStatus
+{
+    if(![self.batteryStatus isBatteryStatusAvailable])
+    {
+        return;
+    }
+    
+    [[NSNotificationCenter defaultCenter] addObserver:self
+                                             selector:@selector(batteryCapacityThresholdDidChange:)
+                                                 name:kKYABatteryCapacityThresholdDidChangeNotification
+                                               object:nil
+     ];
+    
+    // Start receiving battery status changes
+    if([[NSUserDefaults standardUserDefaults] kya_isBatteryCapacityThresholdEnabled])
+    {
+        [self.batteryStatus registerForCapacityChangesIfNeeded];
+    }
+}
+
+- (void)checkAndEnableBatteryOverride
+{
+    CGFloat currentCapacity = self.batteryStatus.currentCapacity;
+    CGFloat threshold = [NSUserDefaults standardUserDefaults].kya_batteryCapacityThreshold;
+    
+    self.batteryOverrideEnabled = (currentCapacity <= threshold);
+}
+
+- (void)disableBatteryOverride
+{
+    self.batteryOverrideEnabled = NO;
+}
+
+- (void)batteryCapacityDidChange:(CGFloat)capacity
+{
+    CGFloat threshold = [NSUserDefaults standardUserDefaults].kya_batteryCapacityThreshold;
+    if([self.sleepWakeTimer isScheduled] && (capacity <= threshold) && ![self isBatteryOverrideEnabled])
+    {
+        [self terminateTimer];
+    }
+}
+
+#pragma mark - Battery Capacity Threshold Changes
+
+- (void)batteryCapacityThresholdDidChange:(NSNotification *)notification
+{
+    if(![self.batteryStatus isBatteryStatusAvailable])
+    {
+        return;
+    }
+    
+    [self terminateTimer];
+}
+
 #pragma mark - Apple Event Manager
 
 - (void)applicationWillFinishLaunching:(NSNotification *)notification
@@ -317,7 +406,6 @@
                                                         andEventID:kAEGetURL
      ];
 }
-
 
 - (void)handleGetURLEvent:(NSAppleEventDescriptor *)event withReplyEvent:(NSAppleEventDescriptor *)reply
 {
@@ -335,10 +423,7 @@
         NSString *minutes = parameters[@"minutes"];
         NSString *hours = parameters[@"hours"];
         
-        if([self.sleepWakeTimer isScheduled])
-        {
-            [self.sleepWakeTimer invalidate];
-        }
+        [self terminateTimer];
         
         // Activate indefinitely if there are no parameters
         if(parameters.count == 0)