+++ KeepingYouAwake/KYABatteryCapacityThreshold.h
@@ -1,0 +1,11 @@
+//
+//  KYABatteryCapacityThreshold.h
+//  KeepingYouAwake
+//
+//  Created by Marcel Dierkes on 01.01.16.
+//  Copyright © 2016 Marcel Dierkes. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+#define kKYABatteryCapacityThresholdDidChangeNotification @"KYABatteryCapacityThresholdDidChangeNotification"