+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 - You can now set the default activation duration for the menu bar icon in preferences
 - Removed the advanced preference for `info.marcel-dierkes.KeepingYouAwake.PreventSleepOnACPower`
 - Added an advanced preference to allow display sleep while still preventing system sleep ([Issue #25](https://github.com/newmarcel/KeepingYouAwake/issues/25))
-- …
+- Ability to set a battery level on MacBooks where the app will deactivate itself ([Issue #24](https://github.com/newmarcel/KeepingYouAwake/issues/24))
+	- *Thank you [timbru31](https://github.com/timbru31) for the suggestion!*
 
 ### v1.3.1 ###